backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.18
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
PyYAML==6.0.1
pyserial==3.5
